fix: Always focus `Rover` when calling `rover.{move|first|last}()`
Behavior before this change:

- Call `rover.first()` (focuses the first `Rover`)
- Move focus elsewhere
- Call `rover.first()` again (focus isn't returned to `Rover`)

This happens because `Rover` is only focused when there's a change on the current selec
ted rover (that is, a change on `rover.currentId`). Since it keeps being the `currentId` when we move focus elsewhere, it doesn't receive focus back when we try to re-select it.

This may be considered a bug as the behavior of `rover.first()` and other move methods
will be inconsistent.

With this change, the behavior is corrected:

- Call `rover.first()` (focuses the first `Rover`)
- Move focus elsewhere
- Call `rover.first()` (focuses the first `Rover`)

To do so, we simply increment a `unstable_moves` state and check if `Rover` is focused
whenever it changes.