fix(Combobox): Fix badly positioned popover on click + minValueLength
Before this commit we were trying to show the Popover as soon as the user click
on the Combobox. This resulted in badly positioned Popover in some cases like
non-static Combobox containers.

Now we check if the Combobox value length is bigger than the minValueLength
before trying to show the Popover.

fixes #808