Try doing a composite project
This is probably not set up correctly and maybe should just be reverted, no idea if this is a viable approach. Notably, testing-library, which successfully supports multiple different languages without having a complicated TS setup, has entirely separate repositories for each of its packages and... that might be a significantly simpler way of doing this but is obviously a pretty big infrastructural change.