Improve support for nested dialogs (#2300)
Closes #2298

This PR adds an example for nested dialogs and fixes two related bugs:

- When the currently active element becomes inaccessible, the focus is
lost to the `document.body` element. Now we're constantly checking
`document.activeElement` to see if the focus was lost, in which case we
move the focus to the dialog element itself.
- When clicking outside a nested modal dialog, it was also closing the
parent dialog. The nested dialog is closed on mousedown, and the parent
dialog is closed on subsequent click event. (#2298)