Support nested/parallel third-party dialogs (#2339)
Closes #2089
Closes #2330
Closes #2232

This PR updates the modal behavior on `Dialog`. Instead of including
only the dialog element in the modal "box", we're now handling the whole
document. This means we can include more elements other than the dialog.

Elements added to the body after the modal dialog is opened will be
automatically marked as nested dialogs. They won't close the underlying
modal dialog when the user interacts with them. This covers most browser
extensions, like Grammarly and Google Translate, and third-party
libraries that only render their dialogs when they're open.

For third-party libraries that render elements early, like
React-Toastify, Intercom, etc., a new `getPersistentElements` function
prop can pass to `Dialog` the elements that should be part of the modal
"box". For non-modal dialogs, like `Popover`, these elements will not
hide the dialog when interacted with.