Add dialog-next-router example (#2369)
This PR adds a new example to the site using Next.js App Router. The
example is presented within an iframe with a chrome bar.

Alongside this, a few adjustments were made to the `Dialog` component:

- Fixed `DialogBackdrop` not including the `data-backdrop` attribute in
the initial render, causing a flash of unstyled content when the dialog
is initially open.
- Fixed `Dialog` calling `hideOnInteractOutside` twice when clicking on
the backdrop.
- The built-in `DialogBackdrop` component is no longer focusable.
- Call `autoFocusOnHide` and `autoFocusOnShow` with a `null` argument
when there's no element to focus or the element is not focusable. This
allows users to specify a fallback element to focus on hide or show.

These changes should make it easier to control the Dialog using the
Next.js App Router.