Render DialogBackdrop as a sibling of Dialog (#2407)
Closes #2085

This PR updates the `Dialog` component to render its backdrop element as
a sibling rather than a parent element. This should make it easier to
perform animations on both the dialog and the backdrop elements
separately.

In addition, we're deprecating the `backdropProps` prop in favor of
passing a JSX element to the `backdrop` prop. This should fix type
errors when using `backdropProps` to pass props to custom backdrop
elements.