Make composite items tabbable before items are rendered (#2601)
This PR updates the logic that controls whether a composite item is
tabbable before hydration. The previous logic was trying to read
`store.item(id)`, which could be populated on the initial render if the
developer is passing an `items` or `defaultItems` prop to the composite
store. Instead, we'll read `state.renderedItems.length`, which will only
be truthy after the items are rendered and registered to the store.