Support dynamic `store` prop on component stores (#2714)
This PR introduces the ability for the dynamic `store` prop on component
store hooks to change between renders. In other words, one store can be
passed to another store using the `store` prop to keep them in sync.
While this functionality is already supported, this PR enhances its
adaptability between renders. This enhancement should simplify the
implementation of #2658.