Throw when passing conflicting props to store (#2745)
This PR introduces an error that's triggered in the following scenario:

```js
const a = useDialogStore();
const b = useDialogStore({ store: a, defaultOpen: true });
```

Both `a` and `b` ought to be synchronized. By default, the last store in
the call stack takes precedence. However, it's still unclear what the
value of the `open` state should be in the example above.

`a` has the `open` state set to `false` by default, while `b` has a
default open state set to `true`. The default state should only be used
when the `open` state is undefined. Since we're passing `a` to `b` and
`a` has the `open` state defined, the `defaultOpen` value is
disregarded, resulting in both `a` and `b` having the `open` state set
to `false`.

This could be confusing, and we might change this behavior in the
future. For now, we're throwing an error to prevent users from relying
on any specific behavior until we make a decision.

_**If you have a particular use case that prefers one behavior over
another, please [submit a feature
request](https://github.com/ariakit/ariakit/issues/new?assignees=&labels=feature&projects=&template=feature.yml).**_