Return composed store props on component stores (#2783)
This PR modifies the component stores that sync with other stores, such
as `useMenuStore` and `useSelectStore`, which accept a `combobox` prop.
Now, they return the provided store as a property of the store object.

For instance:

```js
const combobox = useComboboxStore();
const select = useSelectStore({ combobox });

console.log(combobox === select.combobox); // true
```

This modification enables us to verify in a Select component if it's
being combined with a combobox. We previously relied on a `combobox`
state for this, but it proved to be an inadequate general solution.

Furthermore, I've discontinued support for the `menu` and `select` props
on `useComboboxStore`. These weren't in use anywhere, and they're more
effectively supported as the `combobox` prop on `useSelectStore` and
`useMenuStore`. This was merely a redundancy.