Accept null value on composed store props (#2796)
Allowing `null` as a value for composed store properties makes it easier
for users to override a default value inferred from context. For
instance, if the tree is rendered as `Menu` > `Dialog` > `Menu`, the
last `Menu` could mistakenly recognize the first one as its parent. In
such a situation, the user can pass `parent: null` to `useMenuStore` or
`MenuProvider` to turn off the default value.