Set disclosure/anchor element upon interaction (#3058)
This PR updates the `setAnchorElement` and `setDisclosureElement` on
`Disclosure` and `Popover` components so they're only called upon
interaction with the disclosure/anchor element. This should enable
support for multiple disclosure/anchor elements for the same content
element with more interactions rather than just clicks.

In addition to that, a small update was made to `MenuBar`: adjacent
`Menu` popovers will now open when the focus moves through menu items in
a menubar. Before, they would only open when another menu was already
visible in the menubar.