Revert Menubar always expanding with focus (#3098)
This PR reverts a change introduced in v0.3.6. In that version, the
`Menubar` component was updated so that hovering or focusing on one of
its menu buttons would invariably open the menu popover if another menu
button was already in focus.

This behavior proved to be inconsistent. So, I'm now reverting to the
previous state where menus only expand if another menu is already
expanded.