Improve performance on DialogBackdrop (#4144)
Related to #4075

Although this doesn't fully address the performance issues mentioned in
that post, it's probably better to use `useEffect` here anyway. There
may be a frame where the backdrop is not visible or is behind some
elements if `z-index` values differ, but in that case, users can set a
static `z-index` for both the dialog and the backdrop element.

We should continue measuring to see what else we can do.