Fix `collectionStore.item()` getting out of sync with current items (#4202)
Currently, calling `store.item(id)` may return an item that has been
previously removed from the collection/composite store. This occurs
because we were searching for the item in the public store, which may
lag behind the private store where items are registered synchronously
one by one (this prevents multiple unnecessary updates when rendering
several items).

This PR modifies the `item` method logic to use the private store
instead of the batched public store.

I also updated the tabbable logic introduced in #4191 to consider only
`item?.element` instead of `item?.element?.isConnected`. This is because
when `isConnected` changes, it doesn't always trigger the
`useStoreState` callback, as the component may not re-render or the
state may not update in response to this DOM change.