Initial release of Ariakit Solid (#4116)
**Ariakit Solid tracking issue:**
https://github.com/ariakit/ariakit/issues/4117

---

This is an initial implementation of the Solid version of Ariakit.

The goal is to create a good base to, later on, build the larger
components and features on top. It handles boilerplate, integration with
the monorepo, base features (like the `render` and `wrapElement` props),
some basic components, and changelog/release management. The merging of
this PR will be followed up by the first release of these two packages
(`@ariakit/solid-core` and `@ariakit/solid`).

## Playground sandbox

This is a Stackblitz playground that I keep synced to this PR to test
stuff easily for lack of tests support:
https://stackblitz.com/~/github.com/DaniGuardiola/ariakit-solid-playground?file=src/App.tsx

## Components

A list of components that have been implemented in this PR.

- focus-trap
- group
- heading
- role
- separator
- visually-hidden

## Base features

Foundational features that have been implemented in this PR

- "system": `createHook` and `createInstance`
- types: `Options`, `HTMLProps`, `Props`, etc
- `render` prop
	- "clone element" API (`<As>` component)
	- function API
- `wrapElement` prop
- "hooks" pattern
  - `withOptions`
  - `stableAccessor`
- other utilities like `useId` and `useTagName`

## Initial release prerequisites

~Pending tasks required to merge this PR and release the initial
versions of the packages.~ All done!

- [x] Make TypeScript happy
- [x] Build Solid code
- [x] Fix remaining CI issues
- [x] Take care of READMEs, etc
- [x] Set up changelogs
- [x] Set up releases

## Dependency/progress map

I made the following graph that represents the dependencies between
components. Some notes:

- The "See TODO" annotations are temporary, in the future those will be
replaced with references to other views since it's impossible to
represent all components and their dependencies in a single view. For
example, it might say "See #​6" referring to view number 6.
- Components already implemented are colored green. I keep the image
up-to-date in this PR, so it's a good way to understand the progress at
a glance.


![image](https://github.com/user-attachments/assets/1f143a9e-0dd2-4fe9-861e-3c4b679f9e36)