[Solid] Add support for examples and tests (#4400)
It can probably be improved in the future, but this PR adds a
rudimentary implementation of Ariakit Solid testing. It also introduces
the idea of "unit tests" as a separate concept from "example tests",
since those present a few issues regarding the Ariakit Solid porting
efforts:

- Most example tests depend on other components rather than being
focused on the component at hand, even when not essential to the testing
itself. This makes sense when considering that the rendered content has
a double purpose: to be an example and a fixture for tests. However,
while more components will become available as the porting efforts
progress, many are not available yet which results in tests that are not
usable right away.
- For the same reason, fixtures are more elaborate (e.g. they contain
headers, styles, and other elements irrelevant to the test), which means
more noise. Focused fixtures are better for extensive unit testing.
- Example tests are very useful, however they can sometimes lack a focus
on direct feature testing. I think focused unit tests are necessary and
complementary to example tests.

The `<TestCase>` thing is an attempt at finding a way to provide
multiple, focused fixtures in a single file that strikes a good balance
between hacky and useful. I'm pretty happy with it but feedback is
welcome.

Also, I documented the naming conventions on which the implementation is
based. The idea would be to eventually migrate from `index.tsx` to
`index.react.tsx`, but there's a fallback so everything will still work
for now.

Some final notes:

- Solid testing won't automatically happen for now, which is fine for
porting purposes, for the time being.
- I haven't touched anything in the website code, so I'm unsure if this
will have any effect there.