Website: support `.solid.` and `.react.` in examples/previews (#4405)
Following #4400 up:

## Support the `react` suffix in examples.

To test, rename:

```diff
-examples/tooltip-label/index.tsx
+examples/tooltip-label/index.react.tsx
```

Then open http://localhost:3000/examples/tooltip-label

Also useful to open http://localhost:3000/examples/tooltip for
comparison. ~An important observation is that the new filename is
reflected in the example file tab:~ This is solved now.


![image](https://github.com/user-attachments/assets/39a2cfbf-770f-40c1-bf36-e6255a420fb0)

This is also the case in Stackblitz projects created from the example,
which will contain that file with the `.react` suffix as well. Here's an
instance of this:
https://stackblitz.com/edit/jyw5zyks?file=index.html&theme=dark

### To do

This PR should be followed up by #4404 - which should then:

- [ ] remove the fallback in the `website/build-pages/const.js` regexp:

```diff
-(\.(react|solid))?
+(\.(react|solid))
```

Additionally, in this PR we still need to:

- [x] handle doing rewrites to hide `.react.`.

## Load Solid previews

Solid previews are now loaded through `/examples/` and through
`/previews/` **if a `__solid` suffix is added to the URL** (and if the
Solid version exists). To try this:
http://localhost:3000/examples/separator__solid