Browser tests: add support for other frameworks (#4410)
Instead of hardcoding the preview URL, we use a `preview` function that
allows routing to the appropriate URL depending on the loader, similar
to how we handle unit tests.

Remaining things to figure out / discuss:

- [x] How to skip tests if example doesn't exist for given loader -
ideas: filesystem lookup, check for 404
- [x] Do we also route the tests inside of `website/` (instead of
`examples`)?

---

Update: I took care of both. Details:

- Tests are skipped by doing a filesystem lookup for an example with the
proper loader (e.g. `index.solid.tsx`), similar to how it's done in unit
tests.
- In website examples, it's always assumed that they are React examples
for now. When the Astro rewrite is worked on, a clean approach to
loaders should probably be considered from the start.
- Website tests are also routed with the test fixture. The only
exception is the Ariakit Plus test, since it's very different and does
not test an example. The default playwright `test` function is still
imported there.
- There are 2 tests where, on top of the fixture, there's manual routing
going on for various reasons. The tests are `dialog-details` and
`toolbar` (both in `examples/`). While using the fixture does introduce
the slight overhead of having to navigate to the example path first, and
then navigating again during the test's runtime, it doesn't get in the
way of the test itself. So I think it's fine. These use the `preview()`
utility directly to route to the loader-appropriate URL.