Add @ariakit/tailwind package (#4402)
This PR introduces the new `@ariakit/tailwind` package, implementing the
[Ariakit
Styles](https://gist.github.com/diegohaz/aef9925b1595d83f7e7f0f9e4a9f03ac)
concept as a Tailwind plugin. Please note that this experimental version
is in the early stages of development.

I've also introduced an initial Astro setup to test the integration with
Tailwind v4, which will serve as the foundation for the site rewrite, as
the rest of the project remains on Tailwind v3.