Improve `PopoverArrow` SVG and add `borderWidth` features (#4531)
This pull request includes several enhancements and improvements to the
`PopoverArrow` component in the `@ariakit/react-core` and
`@ariakit/react` packages. The main changes include adding a new prop,
improving the component's functionality, and refactoring the code for
better maintainability.

Enhancements to `PopoverArrow` component:

* Added a new `borderWidth` prop to the `PopoverArrow` component to
allow users to define the border width explicitly.
* Improved the `PopoverArrow` component to infer its border width from
the popover’s `box-shadow` style when applicable, ensuring a consistent
size regardless of the arrow's size.
* Modified the arrow’s SVG path to be more angled in the pointing
direction for better visual accuracy.

Code refactoring and improvements:

* Refactored the `usePopoverArrow` hook to include the new `borderWidth`
prop and improve the logic for determining the border width.
[[1]](diffhunk://#diff-6995e34e35c8bc2bd92fb9642b82052d29e7f6b4c4778debdb5be8d64a61a1a3L52-R70)
[[2]](diffhunk://#diff-6995e34e35c8bc2bd92fb9642b82052d29e7f6b4c4778debdb5be8d64a61a1a3L62-R122)
[[3]](diffhunk://#diff-6995e34e35c8bc2bd92fb9642b82052d29e7f6b4c4778debdb5be8d64a61a1a3L91-R137)
* Updated the `PopoverArrow` component to use `memo` for performance
optimization and refactored the code to use `useStoreState` instead of
`store.useState`.
[[1]](diffhunk://#diff-6995e34e35c8bc2bd92fb9642b82052d29e7f6b4c4778debdb5be8d64a61a1a3L4-R12)
[[2]](diffhunk://#diff-6995e34e35c8bc2bd92fb9642b82052d29e7f6b4c4778debdb5be8d64a61a1a3L28-R34)
[[3]](diffhunk://#diff-6995e34e35c8bc2bd92fb9642b82052d29e7f6b4c4778debdb5be8d64a61a1a3R44-R50)
[[4]](diffhunk://#diff-6995e34e35c8bc2bd92fb9642b82052d29e7f6b4c4778debdb5be8d64a61a1a3L123-R167)
[[5]](diffhunk://#diff-6995e34e35c8bc2bd92fb9642b82052d29e7f6b4c4778debdb5be8d64a61a1a3R188-R192)